+++ NEWS.md
@@ -1,4 +1,5 @@
 ## cmdlR v0.0.1
+* Added a simple search for starting values function
 * Added functions to calculate the standard errors of variances, covariances, correlations and standard deviations of variance-covariance matrix of random parameters in a MIXL model
 * Can specify paramters to stay fixed at the starting values using model_opt$fixed
 * Added convenient function store() governed by the list save_opt() to save results and model objects